chore(misc): re-enable convert-to-monorepo test to test without Nx plugins for now
